test: add export render benchmarks via a crate lib + bin split
Split the binary-only `ajh-tauri` crate into a library (`src/lib.rs`,
holding the whole app plus a `run()` entry) and a thin binary launcher
(`src/main.rs` -> `ajh_tauri::run()`). This is the canonical Tauri 2
layout and makes internal functions reachable from a separate bench
target, which was previously impossible: a bin-only crate exposes no
library API for benches or integration tests. The `windows_subsystem`
attribute stays on the binary; the crate-level await-holding-lock deny
moves to the library root. No runtime behaviour changes.

Add criterion benches (`benches/export_render.rs`) over the export hot
path, the request-to-bytes render pipeline behind documents_export_document:
PDF (single- and two-column templates) and DOCX. They reach the engine
through the public crate API and inline a representative resume fixture
(the in-crate test fixtures are cfg(test) and unreachable from a bench).

Track results over time with a new advisory `benchmark.yml` workflow: it
records a baseline when landing on main (history on a gh-pages branch) and
comments on regressions in pull requests. It never fails the build until
thresholds are tuned, matching the other advisory quality checks.

The lib conversion turns the pipeline builder's `add` method into public
API, which clippy flags as should_implement_trait. It is an intentional
fluent builder verb, so it carries a scoped allow with a reason comment.

Classify `lib.rs` as L3 shell in the architecture boundary tests (and
docs/architecture-rules.md) alongside `main` — it now owns the Tauri
builder, so it is the shell entry point and is exempt from the no-Tauri
rule for lower layers.
